PATCH isolate client header logic and fix mobile drawer accessibility semantics
